add logger interface, move signatures' amount to float32

+++ internal/domain/service/balance.go
@@ -11,8 +11,8 @@ type BalanceStorage interface {
 	// GetAll(ctx context.Context, limit, offset int) ([]entity.Balance, error)
 	// Create(ctx context.Context, balance entity.Balance) (entity.Balance, error)
 	// Update(ctx context.Context, id string, amount int64) error
-	IncreaseAmount(ctx context.Context, id string, amount int64) error
-	DecreaseAmount(ctx context.Context, id string, amount int64) error
+	IncreaseAmount(ctx context.Context, id string, amount float32) error
+	DecreaseAmount(ctx context.Context, id string, amount float32) error
 }
 
 type BalanceService struct {
@@ -27,9 +27,9 @@ func (s BalanceService) GetByID(ctx context.Context, id string) (*entity.Balance
 //	return s.storage.Create(ctx, balance)
 // }
 
-func (s BalanceService) ChangeAmount(ctx context.Context, id string, amount int64) error {
+func (s BalanceService) ChangeAmount(ctx context.Context, id string, amount float32) error {
 	if amount == 0 {
-		return fmt.Errorf("changing balance with id = %s by zero(amount = %d)", id, amount)
+		return fmt.Errorf("changing balance with id = %s by zero(amount = %f)", id, amount)
 	} else if amount > 0 {
 		return s.storage.IncreaseAmount(ctx, id, amount)
 	}
